feat(prism): extract as standalone design token package
Move color tokens, glass utilities, contrast tools, and seasonal
palettes from libs/foliage into libs/prism (@autumnsgrove/prism).
Foliage and engine now import from prism directly. Remove unused
engine foliage re-export barrels and @autumnsgrove/foliage dep.

Closes #1419
