--- libs/engine/src/lib/foliage/components.ts
@@ -1,1 +1,0 @@
-export * from "@autumnsgrove/foliage/components";
