--- libs/engine/src/lib/foliage/index.ts
@@ -1,1 +1,0 @@
-export * from "@autumnsgrove/foliage";
