--- libs/engine/src/lib/foliage/server.ts
@@ -1,1 +1,0 @@
-export * from "@autumnsgrove/foliage/server";
