--- libs/engine/src/lib/foliage/themes.ts
@@ -1,1 +1,0 @@
-export * from "@autumnsgrove/foliage/themes";
