+++ libs/engine/src/routes/vineyard/+page.svelte
@@ -59,7 +59,7 @@
 		midnightBloom,
 		flowers,
 	} from "$lib/ui/components/nature/palette";
-	import { groveColors as grove, cream, bark as barkTokens, status } from "@autumnsgrove/foliage";
+	import { groveColors as grove, cream, bark as barkTokens, status } from "@autumnsgrove/prism";
 	import { Sparkles, Palette, Box, ChevronRight, Type } from "lucide-svelte";
 
 	// Interactive state for demos
